+++ mimesis/data/int/__init__.py
@@ -3,12 +3,12 @@
 """Here you can see data that used in all locales."""
 
 from .address import *
-from .business import *
 from .code import *
 from .datetime import *
 from .decorator import *
 from .development import *
 from .file import *
+from .finance import *
 from .hardware import *
 from .internet import *
 from .path import *