+++ mimesis/providers/__init__.py
@@ -5,7 +5,6 @@
 from mimesis.providers.address import Address
 from mimesis.providers.base import BaseDataProvider, BaseProvider
 from mimesis.providers.binaryfile import BinaryFile
-from mimesis.providers.business import Business
 from mimesis.providers.choice import Choice
 from mimesis.providers.clothing import Clothing
 from mimesis.providers.code import Code
@@ -13,6 +12,7 @@ from mimesis.providers.cryptographic import Cryptographic
 from mimesis.providers.date import Datetime
 from mimesis.providers.development import Development
 from mimesis.providers.file import File
+from mimesis.providers.finance import Finance
 from mimesis.providers.food import Food
 from mimesis.providers.generic import Generic
 from mimesis.providers.hardware import Hardware
@@ -33,7 +33,7 @@ __all__ = [
     # Data providers:
     "Address",
     "BinaryFile",
-    "Business",
+    "Finance",
     "Choice",
     "Clothing",
     "Code",