+++ tests/test_providers/patterns.py
@@ -25,13 +25,13 @@ EMAIL_REGEX = r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)"
 CREDIT_CARD_REGEX = r"[\d]+((-|\s)?[\d]+)+"
 
 PROVIDER_STR_REGEX = (
-    r"^(Business|Clothing|Code|Development"
+    r"^(Finance|Clothing|Code|Development"
     r"|File|Games|Hardware|Internet|Numbers|Path|Payment|"
     r"Transport|Cryptographic|Science)"
 )
 
 DATA_PROVIDER_STR_REGEX = (
-    r"^(Address|Business|Datetime|Food|"
+    r"^(Address|Finance|Datetime|Food|"
     r"Person|Text|BaseDataProvider|BaseField|Field) <([a-z-]+)>$"
 )